Attempted to set up form for the student account creation and passing data to the backend
