+++ Project/student_account/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class StudentAccountConfig(AppConfig):
+    name = 'student_account'
