+++ Project/student_account/views.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def index(request):
+    return render(request, 'student_account/student_account.html')
