+++ Project/templates/contact_page/contact.html
@@ -8,8 +8,12 @@
     <div class="container" style="max-width: 500px; padding: 10px; background-color: #d6d8d9; margin: 0 auto; min-height: 100%; overflow: hidden">
         <form action="/action_page.php" class="md-form" >
             <div class="form-group">
-                <label for="FullName">Full name: *</label>
-                <input type="text" class="form-control" id="FullName" required>
+            <label for="FirstName">First Name: *</label>
+            <input type="text" class="form-control" id="FirstName" required>
+            </div>
+            <div class="form-group">
+            <label for="LastName">Last Name: *</label>
+            <input type="text" class="form-control" id="LastName" required>
             </div>
             <div class="form-group">
                 <label for="DoB">Date of Birth: *</label>
