+++ Project/templates/contact_page/includes/teach_contact.html
@@ -1,4 +1,4 @@
 {% block content %}
-
+Teacher contact etc
 
 {% endblock %}
