+++ Project/templates/header.html
@@ -20,7 +20,7 @@
 <body class="body" style="background-color:#f6f6f6">
 	<div class="container-fluid" style="min-height:95%; ">
 		<div class="row">
-			  <div class="col-sm-2">
+			  <div class="col-sm-1">
 				  <br>
 				  <center>
 				  </center>
@@ -32,7 +32,7 @@
 				  </center>
 			  </div>
 		</div><hr>
-		
+
 
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
   <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
@@ -54,7 +54,10 @@
       </li>
       <li class="nav-item">
          <a class="nav-link" href='/careers/'>Careers</a>
-       </li>
+      </li>
+      <li class="nav-item">
+          <a class="nav-link" href='/create_account/'>Create Student Account</a>
+      </li>
         {% if request.user.is_authenticated %}
             <li class="nav-item">
                 <a class="nav-link" href='/account/'>Account</a>
@@ -80,16 +83,15 @@
   </div>
 </nav>
 
-        <div class="row">
-		  <div class="col-sm-10">
-		  
-			<div class='container-fluid'>
-			<br><br>
-			   {% block content %}
-			   {% endblock %}	
-			</div>
-		  </div>
-		</div> 
+
+
+    <div class='container-fluid'>
+    <br><br>
+       {% block content %}
+       {% endblock %}
+    </div>
+
+
 	</div>
 	<footer>
 	</footer>
