+++ Project/templates/student_account/student_account.html
@@ -1,0 +1,67 @@
+{% extends "header.html" %}
+
+{% block content %}
+
+
+<!-- Control the column width, and how they should appear on different devices -->
+<!--<form method="post">-->
+
+    {% csrf_token %}
+    {{ form.as_p }}
+    <div class="container" style="max-width: 1000px; padding: 10px; background-color: #d6d8d9; margin: 0 auto; min-height: 100%; overflow: hidden">
+    <form method=post action="/action_page.php" class="md-form" >
+        <div class="form-row">
+            <label for="LastName">First Name: *</label>
+        <div class="col">
+            <input type="text" class="form-control" name = "FirstName" id="FirstName"  placeholder="First name">
+        </div>
+            <label for="FirstName">Last Name: *</label>
+        <div class="col">
+            <input type="text" class="form-control" name = "LastName" id="LastName" placeholder="Last name">
+        </div>
+        </div>
+        <div class="form-group">
+            <label for="DoB">Date of Birth: *</label>
+            <input type="date" class="form-control" name="DoB" id="DoB" required>
+        </div>
+        <div class="form-check form-check-inline">
+            <input class="form-check-input" type="radio" name="inlineRadioOptions" id="Male" value="option1">
+            <label class="form-check-label" for="Male">Male</label>
+        </div>
+        <div class="form-check form-check-inline">
+            <input class="form-check-input" type="radio" name="inlineRadioOptions" id="Female" value="option2">
+            <label class="form-check-label" for="Female">Female</label>
+        </div>
+        <div class="form-check form-check-inline">
+            <input class="form-check-input" type="radio" name="inlineRadioOptions" id="Other" value="option3">
+            <label class="form-check-label" for="Other">Other</label>
+        </div>
+        <div class="form-group">
+            <label for="HomeAddress">Home address:</label>
+            <input type="text" class="form-control" id="HomeAddress">
+        </div>
+        <div class="form-group">
+            <label for="email">Email address: *</label>
+            <input type="email" class="form-control" id="email" required>
+        </div>
+        <div class="form-group">
+            <label for="pNumber">Phone Number: *</label>
+            <input type="text" class="form-control" id="pNumber" required>
+        </div>
+        <div class="form-group">
+            <label for="certificates">Qualifications: * (e.g. certificates, experience)</label>
+            <textarea class="form-control" rows="4" id="certificates" required></textarea>
+        </div>
+        <div class="form-group">
+            <label for="subjects">What instruments are you interrested in learning?</label>
+            <textarea class="form-control" rows="2" id="subjects" ></textarea>
+        </div>
+        <div class="form-group">
+            <button type="submit" style="float: right;" class="btn btn-primary">Submit</button>
+        </div>
+    <!--</form>-->
+    </div>
+<!--</form>-->
+<!--</div>-->
+
+{% endblock %}
