+++ Project/tests/test_welcome_page.py
@@ -1,34 +1,39 @@
-# import unittest
-# from django.http import HttpRequest
-# from django.test import SimpleTestCase
-# from django.urls import reverse
-#
-#
-# class HomePageTests(SimpleTestCase):
-#
-#     def test_home_page_status_code(self):
-#         response = self.client.get('/')
-#         self.assertEquals(response.status_code, 200)
-#
-#     def test_view_url_by_name(self):
-#         response = self.client.get(reverse('home'))
-#         self.assertEquals(response.status_code, 200)
-#
-#     def test_view_uses_correct_template(self):
-#         response = self.client.get(reverse('home'))
-#         self.assertEquals(response.status_code, 200)
-#         self.assertTemplateUsed(response, 'header.html')
-#
-#     def test_home_page_contains_correct_html(self):
-#         response = self.client.get('/')
-#         self.assertContains(response, 'Welcome to Pinelands Music School')
-#
-#     def test_home_page_does_not_contain_incorrect_html(self):
-#         response = self.client.get('/')
-#         self.assertNotContains(
-#             response, 'Hi there! I should not be on the page.')
-#
-#
-#
-# if __name__ == '__main__':
-#     unittest.main()
+import unittest
+from django.http import HttpRequest
+from django.test import SimpleTestCase
+from django.urls import reverse
+
+
+class HomePageTests(SimpleTestCase):
+
+    def test_home_page_status_code(self):
+        response = self.client.get('/')
+        self.assertEquals(response.status_code, 200)
+
+    def test_view_url_by_name(self):
+        response = self.client.get(reverse('home'))
+        self.assertEquals(response.status_code, 200)
+
+    def test_view_uses_correct_template(self):
+        response = self.client.get(reverse('home'))
+        self.assertEquals(response.status_code, 200)
+        self.assertTemplateUsed(response, 'header.html')
+
+    def test_home_page_contains_correct_html(self):
+        response = self.client.get('/')
+        self.assertContains(response, 'Welcome to Pinelands Music School')
+
+    def test_home_page_does_not_contain_incorrect_html(self):
+        response = self.client.get('/')
+        self.assertNotContains(
+            response, 'Hi there! I should not be on the page.')
+
+    def test_link_url_by_name(self):
+        url = "/create_account/"
+        response = self.client.get(reverse('url'))
+        self.assertEquals(response.status_code, 200)
+
+
+
+if __name__ == '__main__':
+    unittest.main()
